requirements.txt
PyYAML==6.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
matplotlib==3.5.3
gym==0.26.2
tensorboard==2.11.2